% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er-functions.R
\name{choose.p.lwo}
\alias{choose.p.lwo}
\title{Choose Specific Entropy Rate Order via Leave-Window-Out Cross-validation}
\usage{
choose.p.lwo(x, p.max, p.min = 1, half.window.length = 1,
  kerntype = "gaussian", bwtype = "fixed", bwmethod = "cv.ml",
  bwuse = NA, nmulti = 1)
}
\arguments{
\item{x}{A continuous-valued scalar time series.}

\item{p.max}{The largest model order to consider.}

\item{p.min}{The smallest model order to consider.}

\item{half.window.length}{The half-length of the window about the present removed in estimating the specific entropy rate.}

\item{kerntype}{The kernel type, one of {'guassian', 'uniform', 'epanechnikov'}.}

\item{bwtype}{The bandwidth type, one of {'fixed', 'generalized_nn', 'adaptive_nn'}.}

\item{bwmethod}{The method used to choose the bandwidths, one of {cv.ml, cv.ls, normal-reference}. Defaults to cv.ml to perform LOOCV on the specific entropy rate.}

\item{bwuse}{Specify a fixed bandwidth to use across all lags, or NA to estimate the bandwidths in a data-driven way.}

\item{nmulti}{The number of random initial conditions to use by the optimizer for choosing the bandwidths.}
}
\description{
choose.p.lwo chooses the model order p to use in estimating the specific entropy
rate of a scalar, continuous-valued time series using leave-window-out 
(LWO) cross validation.
}

