% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-functions.R
\name{embed.ts}
\alias{embed.ts}
\title{Embed a Scalar Time Series}
\usage{
embed.ts(ts, L = 2)
}
\arguments{
\item{ts}{The scalar time series.}

\item{L}{The block length used to embed.}
}
\description{
Embeds a scalar time series in a regression-like matrix.
}
\keyword{embedding,}
\keyword{regression}
\keyword{series,}
\keyword{time}

