% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er-functions.R
\name{estimate.pred.dens}
\alias{estimate.pred.dens}
\title{Estimate Predictive / Conditional Density}
\usage{
estimate.pred.dens(x, p, kerntype = "gaussian", bwtype = "fixed",
  bwmethod = "cv.ml", bwuse = NA, nmulti = 1)
}
\arguments{
\item{x}{A continuous-valued scalar time series.}

\item{p}{The order of the conditional density.}

\item{kerntype}{The kernel type, one of {'guassian', 'uniform', 'epanechnikov'}.}

\item{bwtype}{The bandwidth type, one of {'fixed', 'generalized_nn', 'adaptive_nn'}.}

\item{bwmethod}{The method used to choose the bandwidths, one of {cv.ml, cv.ls, normal-reference}. Defaults to cv.ml to perform LOOCV on the specific entropy rate.}

\item{bwuse}{Specify a fixed bandwidth to use across all lags, or NA to estimate the bandwidths in a data-driven way.}

\item{nmulti}{The number of random initial conditions to use by the optimizer for choosing the bandwidths.}
}
\description{
estimate.pred.dens estimates the predictive / conditional density of
order p associated with the scalar time series x using np.
}
\keyword{conditional}
\keyword{density}
\keyword{density,}
\keyword{estimator}
\keyword{kernel}
\keyword{predictive}

