% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-functions.R
\name{gaussian.kernel}
\alias{gaussian.kernel}
\title{Gaussian Kernel of Scalar}
\usage{
gaussian.kernel(u)
}
\arguments{
\item{u}{Evaluation points}
}
\description{
Returns the standard Gaussian kernel evaluated at u.
}
\keyword{kernel}

