% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-functions.R
\name{uniformize}
\alias{uniformize}
\title{Transform to Uniform Scale}
\usage{
uniformize(x)
}
\arguments{
\item{x}{A vector of values.}
}
\description{
Transform a vector with support on [a, b] to have
support on [0, 1].
}

